+++ include/habitify_event_bus/publisher.h
@@ -17,7 +17,7 @@
 // Contact via <https://github.com/SPauly/habitify-event-bus>
 #ifndef HABITIFY_EVENT_BUS_INCLUDE_PUBLISHER_H_
 #define HABITIFY_EVENT_BUS_INCLUDE_PUBLISHER_H_
-#include <spauly/habitify_event_bus/impl/publisher_base.h>
+#include <habitify_event_bus/impl/publisher_base.h>
 
 #include <cassert>
 #include <condition_variable>
@@ -25,9 +25,9 @@
 #include <shared_mutex>
 #include <vector>
 
-#include <spauly/habitify_event_bus/actor_ids.h>
-#include <spauly/habitify_event_bus/event.h>
-#include <spauly/habitify_event_bus/impl/port.h>
+#include <habitify_event_bus/actor_ids.h>
+#include <habitify_event_bus/event.h>
+#include <habitify_event_bus/impl/port.h>
 
 namespace habitify_event_bus {
 class EventBusImpl;
